+++ script/SafeConfig.sol
@@ -7,9 +7,6 @@ library SafeConfig {
     function _isTestnet() internal view returns (bool) {
         if (
             block.chainid == 11155111 // sepolia
-                || block.chainid == 84532 // base sepolia
-                || block.chainid == 421614 // arbitrum sepolia
-                || block.chainid == 11155420 // optimism sepolia
         ) {
             return true;
         }
@@ -18,6 +15,7 @@ library SafeConfig {
                 || block.chainid == 10 // optimism
                 || block.chainid == 56 // bnb
                 || block.chainid == 137 // polygon
+                || block.chainid == 480 // worldchain
                 || block.chainid == 5000 // mantle
                 || block.chainid == 8453 // base
                 || block.chainid == 34443 // mode
@@ -25,6 +23,7 @@ library SafeConfig {
                 || block.chainid == 43114 // avalanche
                 || block.chainid == 59144 // linea
                 || block.chainid == 81457 // blast
+                || block.chainid == 167000 // taiko
                 || block.chainid == 534352 // scroll
         ) {
             return false;
@@ -40,6 +39,7 @@ library SafeConfig {
             block.chainid == 10 // optimism
                 || block.chainid == 56 // bnb
                 || block.chainid == 137 // polygon
+                || block.chainid == 480 // worldchain
                 || block.chainid == 5000 // mantle
                 || block.chainid == 8453 // base
                 || block.chainid == 34443 // mode
@@ -47,11 +47,9 @@ library SafeConfig {
                 || block.chainid == 43114 // avalanche
                 || block.chainid == 59144 // linea
                 || block.chainid == 81457 // blast
-                || block.chainid == 84532 // base sepolia
-                || block.chainid == 421614 // arbitrum sepolia
+                || block.chainid == 167000 // taiko
                 || block.chainid == 534352 // scroll
                 || block.chainid == 11155111 // sepolia
-                || block.chainid == 11155420 // optimism sepolia
         ) {
             return false;
         }