+++ sh/common.sh
@@ -7,9 +7,9 @@ declare forge_version
 forge_version="$(forge --version)"
 forge_version="${forge_version:13:7}"
 declare -r forge_version
-if [[ $forge_version != 'fe2acca' ]] ; then
+if [[ $forge_version != '59f354c' ]] ; then
     echo 'Wrong foundry version installed -- '"$forge_version" >&2
-    echo 'Run `foundryup -v nightly-fe2acca4e379793539db80e032d76ffe0110298b`' >&2
+    echo 'Run `foundryup -v nightly-59f354c179f4e7f6d7292acb3d068815c79286d1`' >&2
     exit 1
 fi
 
@@ -59,3 +59,34 @@ if [[ $(get_config isCancun) != [Tt]rue ]] ; then
     echo 'You are on the wrong branch' >&2
     exit 1
 fi
+
+declare -i chainid
+chainid="$(get_config chainId)"
+declare -r -i chainid
+
+declare rpc_url
+rpc_url="$(get_api_secret rpcUrl)"
+declare -r rpc_url
+
+if [[ ${rpc_url:-unset} = 'unset' ]] || [[ $rpc_url == 'null' ]] ; then
+    echo '`rpcUrl` is unset in `api_secrets.json` for chain "'"$chain_name"'"' >&2
+    exit 1
+fi
+
+function verify_contract {
+    declare -r _verify_constructor_args="$1"
+    shift
+    declare -r _verify_deployed_address="$1"
+    shift
+    declare -r _verify_source_path="$1"
+    shift
+
+    if (( chainid == 34443 )) ; then # Mode uses Blockscout, not Etherscan
+        forge verify-contract --watch --chain $chainid --verifier blockscout --verifier-url "$(get_config blockscoutApi)" --constructor-args "$_verify_constructor_args" "$_verify_deployed_address" "$_verify_source_path"
+    else
+        forge verify-contract --watch --chain $chainid --verifier etherscan --etherscan-api-key "$(get_api_secret etherscanKey)" --verifier-url "$(get_config etherscanApi)" --constructor-args "$_verify_constructor_args" "$_verify_deployed_address" "$_verify_source_path"
+    fi
+    if (( chainid != 480 )) && (( chainid != 81457 )) && (( chainid != 167000 )); then # Sourcify doesn't support World Chain, Blast, or Taiko
+        forge verify-contract --watch --chain $chainid --verifier sourcify --constructor-args "$_verify_constructor_args" "$_verify_deployed_address" "$_verify_source_path"
+    fi
+}