+++ src/chains/Blast/TakerSubmitted.sol
@@ -1,0 +1,41 @@
+// SPDX-License-Identifier: MIT
+pragma solidity =0.8.25;
+
+import {BlastMixin} from "./Common.sol";
+import {Settler} from "../../Settler.sol";
+
+import {IERC20} from "@forge-std/interfaces/IERC20.sol";
+import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
+import {ISettlerActions} from "../../ISettlerActions.sol";
+
+// Solidity inheritance is stupid
+import {SettlerAbstract} from "../../SettlerAbstract.sol";
+import {SettlerBase} from "../../SettlerBase.sol";
+import {Permit2PaymentAbstract} from "../../core/Permit2PaymentAbstract.sol";
+import {AbstractContext} from "../../Context.sol";
+
+/// @custom:security-contact security@0x.org
+contract BlastSettler is Settler, BlastMixin {
+    constructor(bytes20 gitCommit) SettlerBase(gitCommit) {}
+
+    function _dispatchVIP(uint256 action, bytes calldata data) internal override DANGEROUS_freeMemory returns (bool) {
+        return super._dispatchVIP(action, data);
+    }
+
+    function _isRestrictedTarget(address target) internal pure override(Settler, BlastMixin) returns (bool) {
+        return BlastMixin._isRestrictedTarget(target) || Settler._isRestrictedTarget(target);
+    }
+
+    // Solidity inheritance is stupid
+    function _dispatch(uint256 i, uint256 action, bytes calldata data)
+        internal
+        override(SettlerAbstract, SettlerBase, BlastMixin)
+        returns (bool)
+    {
+        return super._dispatch(i, action, data);
+    }
+
+    function _msgSender() internal view override(Settler, AbstractContext) returns (address) {
+        return super._msgSender();
+    }
+}