+++ CHANGELOG.md
@@ -24,6 +24,8 @@ Master list of UniV3 forks:
   21. SupSwapV3
   22. Kim (Algebra-like)
   23. SwapMode
+  24. Swapsicle (Algebra-like)
+  25. Panko
 
 ---
 
@@ -51,6 +53,13 @@ Master list of UniV3 forks:
   * Add `DODOV1` action to Mantle
   * Add `DODOV1` action to Polygon
   * Add `DODOV1` action to Scroll
+* Deploy Settler to Taiko network
+  * Add UniswapV3 UniV3 fork to Taiko
+  * Add Swapsicle UniV3 (Algebra style) fork to Taiko
+  * Add Panko UniV3 fork to Taiko
+* Deploy Settler to World Chain network
+  * Add UniswapV3 UniV3 fork to World Chain
+  * Add DackieSwapV3 UniV3 fork to World Chain
 * Create new Settler version for intent-based settlement -- the taker only signs
   the slippage, not the actions
   * This is now `tokenId` 4